Set Float & Float2 inside sources.py

+++ tests/test_deposit.py
@@ -18,8 +18,8 @@ mass = 1.0
 manifold = Manifold(nx, ny, comm)
 
 # # Initialize sources
-sources = Sources(manifold, dtype=Float)
-sources_ppic2 = Sources(manifold, dtype=Float)
+sources = Sources(manifold)
+sources_ppic2 = Sources(manifold)
 
 # Total number of particles
 np = npc*nx*ny
